@@ -1,3 +1,4 @@
 #define l_mem ((volatile unsigned int*)0x00000000)
 #define l_mem_size 512
-#define LED1 (*(volatile unsigned int*)0x01000000)
+#define data ((volatile unsigned short*)0x01000000)
+#define data_size 16
